pyproject.toml
[tool.poetry]
name = "credit-card-vc"
version = "0.1.0"
description = "Credit Card Service with FastAPI"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.2"
dependency-injector = "4.44.0"
aws-cdk-lib = "^2.154.0"
constructs = "^10.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
